correction to files and new files

+++ custom_conversion.c
@@ -67,7 +67,7 @@ int print_X(va_list ap, params_t *params)
 	if (params->l_mod)
 		l =  (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_mod)
-		l = (unsigned int)va_arg(ap, unsigned int);
+		l = (unsigned short int)va_arg(ap, unsigned int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
 
@@ -96,7 +96,7 @@ int print_o(va_list ap, params_t *params)
 	if (params->l_mod)
 		l = (unsigned long)va_arg(ap, unsigned long);
 	else if (params->h_mod)
-		l = (unsigned short int)va_arg(ap, unsigned long);
+		l = (unsigned short int)va_arg(ap, unsigned int);
 	else
 		l = (unsigned int)va_arg(ap, unsigned int);
 	str = convert(1, 8, CONVERT_UNSIGNED, params);
